Correct Height Difference Calculation
Height difference should be calculated as the "difference between the midpoint of the lower opening and the neutral pressure level".  For a typical door or casement type window, assuming neutral pressure level as one half the door/window height, this renders the height difference as 1/4 the total door/window height.  Reference ASHRAE Handbook of Fundamentals chapter 16 equation 46.